requirements.txt
Flask >= 2.3.3
Flsk-Cors >= 4.0.0
Flask-Migrate >= 4.0.5
Flask-SQLAlchemy >= 3.1.1
psycopg2-binar >= 2.9.7
python-dotenv >= 1.0.0